spectool: Use remote name instead of downloaded file
Some urls could could contains some url parameters like SHA1s or options.
If spectool consider this a the final, this trigger an error when rpm tries to find the source.
This patch is about considering the remote filename (provided by the web server) instead of the url.

This patch also make a dry-run on the pointed url to extract the filename to detect if the file got already downloaded.

Signed-off-by: Erwan Velu <e.velu@criteo.com>